Sponsor list is now dynamic, stored in /data/_sponsors/
4 variables:

`name:` is the name of the sponsor, this is required.
`importance:` is the ranking of sponsors, this is required.
`image:` is a relevant image of the sponsor, this should only be used on high level sponsors.
`link:` is a link to the sponsor's website.

NAMING SCHEME:

xy-zz-zz-zz.md

x: importance level
y: number within importance level
z: name of sponsor